state: Move initialization code to it's own file
